fix: Limit host connections getting subscription feeds
When attempting to view My Subscriptions some users are reporting they
are unable to see any results with following errors in logs:

    ssl.SSLZeroReturnError: TLS/SSL connection has been closed (EOF) (_ssl.c:997)

This is a result on certain systems of attempting to fetch too many
subscription channels (~50 or more) since the current code generates a
thread and connection for every channel feed.

Fixed using a requests session to reuse the connection and setting a
maximum pool size of 5 with blocking enabled to prevent new connections.
There is also a speedup from limiting the pool size (0.5s vs 5s) and the
selected pool size seems a good compromise.

Fixes: https://github.com/anxdpanic/plugin.video.youtube/issues/280